aiagent/cli_requirements.txt
# CLI Tool Dependencies for ForensIQ
# Add these to your existing requirements.txt or install separately

# Async HTTP client for API communication
aiohttp>=3.8.0

# Async file operations
aiofiles>=23.0.0

# Encryption for secure credential storage
cryptography>=41.0.0

# Task scheduling
schedule>=1.2.0

# HTTP requests (alternative to aiohttp for sync operations)
requests>=2.31.0

# MongoDB drivers for real database integration
motor>=3.3.0
pymongo>=4.5.0

# Environment variable loading
python-dotenv>=1.0.0

# Pydantic for data validation (same as server)
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Command line argument parsing (built-in)
# argparse (standard library)

# JSON handling (built-in)
# json (standard library)

# Password input without echo
# getpass (standard library)

# Path handling (built-in)
# pathlib (standard library)

# Time operations (built-in)
# time (standard library)

# Logging (built-in)
# logging (standard library)

# Base64 encoding (built-in)
# base64 (standard library)

# Hash functions (built-in)
# hashlib (standard library)

# OS operations (built-in)
# os (standard library)

# System operations (built-in)
# sys (standard library)

# Async operations (built-in in Python 3.7+)
# asyncio (standard library)

# Date and time operations (built-in)
# datetime (standard library)

# Type hints (built-in in Python 3.5+)
# typing (standard library)
